+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 0.6.4 ()
 
 - Bugfix #274: `initial` has not been passed to super in `HierachicalMachine.add_model` (thanks @illes)
+- Feature #275: `HierarchicalMachine.add_states` now supports keyword `parent` to be a `NestedState` or a string.
 
 ## 0.6.3 (November, 2017)